requirements.txt
gradio
google-generativeai
flask
